--- data/books.json
@@ -1,52 +1,0 @@
-[
-  {
-    "id": "4d8f6e30-5c7a-40fd-a2a6-aa7932cf2b36",
-    "titulo": "O Pequeno Príncipe",
-    "autor": "Antoine de Saint-Exupéry",
-    "genero": "Fábula",
-    "anoPublicacao": 1943,
-    "paginas": 365,
-    "status": "aberto",
-    "avaliacao": 5
-  },
-  {
-    "id": "54d5b923-711d-4acd-a07d-943c58b2d51d",
-    "titulo": "O Pequeno Príncipe",
-    "autor": "Antoine de Saint-Exupéry",
-    "genero": "Fábula",
-    "anoPublicacao": 1943,
-    "paginas": 365,
-    "status": "finalizado",
-    "avaliacao": 5
-  },
-  {
-    "id": "1789acbe-0b79-47db-a79d-9aa1e115c37a",
-    "titulo": "O Pequeno Príncipe",
-    "autor": "Antoine de Saint-Exupéry",
-    "genero": "Fábula",
-    "anoPublicacao": 1943,
-    "paginas": 365,
-    "status": "fechado",
-    "avaliacao": 0
-  },
-  {
-    "id": "8f52c0d5-187e-4649-8304-95891f231b6f",
-    "titulo": "O Pequeno Príncipe",
-    "autor": "Antoine de Saint-Exupéry",
-    "genero": "Fábula",
-    "anoPublicacao": 1943,
-    "paginas": 365,
-    "status": "fechado",
-    "avaliacao": 0
-  },
-  {
-    "id": "e14afb52-14d8-41ca-9bfa-956c105d3c9f",
-    "titulo": "O Pequeno Príncipe",
-    "autor": "Antoine de Saint-Exupéry",
-    "genero": "Fábula",
-    "anoPublicacao": 1943,
-    "paginas": 365,
-    "status": "finalizado",
-    "avaliacao": 0
-  }
-]
