+++ readfy-app/src/app/page.tsx
@@ -1,51 +1,51 @@
-import { db } from "../lib/db";
-import { computeStats } from "../lib/metrics";
-import { StatCard } from "../components/dashboard/stat-card";
-import { QuickActions } from "../components/dashboard/quick-actions";
+import Link from 'next/link';
 
-export const dynamic = "force-dynamic";
-
-export default async function DashboardPage() {
-  const books = await db.list();
-  const stats = computeStats(books);
-
+export default function Home() {
   return (
-    <section className="space-y-6">
-      <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-2">
-        <h1 className="text-xl sm:text-2xl font-semibold tracking-tight">
-          Dashboard
-        </h1>
-        <QuickActions />
-      </div>
+    <div className="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center p-6">
+      <div className="max-w-4xl mx-auto text-center">
+        <div className="mb-12">
+          <h1 className="text-5xl font-bold text-gray-900 mb-4">
+            📚 Sistema de Gerenciamento de Livros
+          </h1>
+          <p className="text-xl text-gray-600">
+            Organize sua biblioteca pessoal de forma simples e eficiente
+          </p>
+        </div>
 
-      <div className="grid grid-cols-2 gap-3 sm:grid-cols-4 sm:gap-4">
-        {/* Estatísticas*/}
-      </div>
+        <div className="flex flex-col sm:flex-row gap-4 justify-center mb-12">
+          <Link 
+            href="/dashboard"
+            className="bg-blue-500 hover:bg-blue-600 text-white px-8 py-3 rounded-lg font-semibold transition-colors"
+          >
+            📊 Acessar Dashboard
+          </Link>
+          
+          <Link 
+            href="/book/register"
+            className="bg-green-500 hover:bg-green-600 text-white px-8 py-3 rounded-lg font-semibold transition-colors"
+          >
+            ➕ Cadastrar Novo Livro
+          </Link>
+        </div>
 
-      <div className="space-y-3">
-        <h2 className="text-base font-medium text-muted-foreground">
-          Recentes
-        </h2>
-        <ul className="divide-y rounded-lg border overflow-hidden">
-          {books.slice(0, 6).map((b) => (
-            <li
-              key={b.id}
-              className="flex items-center justify-between p-3 sm:p-4"
-            >
-              <div className="min-w-0">
-                <p className="truncate font-medium">{b.titulo}</p>
-                <p className="truncate text-sm text-muted-foreground">
-                  {b.autor}
-                  {b.genero ? ` • ${b.genero}` : ""}
-                </p>
-              </div>
-              <span className="text-xs rounded-full bg-secondary px-2 py-1">
-                {b.status ?? "—"}
-              </span>
-            </li>
-          ))}
-        </ul>
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-6 text-left">
+          <div className="bg-white/50 backdrop-blur-sm rounded-lg p-6">
+            <h4 className="font-semibold mb-2">🔍 Busca Inteligente</h4>
+            <p className="text-sm text-gray-600">Encontre livros por título, autor ou gênero</p>
+          </div>
+          
+          <div className="bg-white/50 backdrop-blur-sm rounded-lg p-6">
+            <h4 className="font-semibold mb-2">📈 Estatísticas</h4>
+            <p className="text-sm text-gray-600">Acompanhe seu progresso de leitura</p>
+          </div>
+          
+          <div className="bg-white/50 backdrop-blur-sm rounded-lg p-6">
+            <h4 className="font-semibold mb-2">⚡ Rápido e Fácil</h4>
+            <p className="text-sm text-gray-600">Interface simples e intuitiva</p>
+          </div>
+        </div>
       </div>
-    </section>
+    </div>
   );
 }
