--- readfy-app/src/components/dashboard/stat-card.tsx
@@ -1,29 +1,0 @@
-import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
-
-export function StatCard({
-  title,
-  value,
-  helper,
-}: {
-  title: string;
-  value: string | number;
-  helper?: string;
-}) {
-  return (
-    <Card>
-      <CardHeader className="pb-2">
-        <CardTitle className="text-base sm:text-sm font-medium text-muted-foreground">
-          {title}
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="pt-0">
-        <div className="text-2xl sm:text-3xl font-semibold tracking-tight">
-          {value}
-        </div>
-        {helper ? (
-          <p className="mt-1 text-xs text-muted-foreground">{helper}</p>
-        ) : null}
-      </CardContent>
-    </Card>
-  );
-}
