--- readfy-app/src/components/theme/theme-toggle.tsx
@@ -1,53 +1,0 @@
-"use client";
-import { useTheme } from "next-themes";
-import { Monitor, Moon, Sun } from "lucide-react";
-import {
-  DropdownMenu,
-  DropdownMenuTrigger,
-  DropdownMenuItem,
-  DropdownMenuContent,
-} from "@/components/ui/dropdown-menu";
-import { Button } from "@/components/ui/button";
-import { useEffect, useState } from "react";
-
-export function ThemeToggle() {
-  const { theme, setTheme, systemTheme } = useTheme();
-  const [mounted, setMounted] = useState(false);
-  useEffect(() => setMounted(true), []);
-  const active = theme === "system" ? systemTheme : theme;
-
-  const icon =
-    active === "dark" ? (
-      <Moon className="size-4" />
-    ) : active === "light" ? (
-      <Sun className="size-4" />
-    ) : (
-      <Monitor className="size-4" />
-    );
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="ghost"
-          size="icon"
-          aria-label="Alternar tema"
-          className="rounded-2xl"
-        >
-          {mounted ? icon : <Monitor className="size-4" />}
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-44">
-        <DropdownMenuItem onClick={() => setTheme("light")} className="gap-2">
-          <Sun className="size-4" /> Claro
-        </DropdownMenuItem>
-        <DropdownMenuItem onClick={() => setTheme("dark")} className="gap-2">
-          <Moon className="size-4" /> Escuro
-        </DropdownMenuItem>
-        <DropdownMenuItem onClick={() => setTheme("system")} className="gap-2">
-          <Monitor className="size-4" /> Sistema
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
