--- readfy-app/src/lib/metrics.ts
@@ -1,5 +1,0 @@
-import { Book } from "./../app/types/book";
-
-export function computeStats(books: Book[]) {
-  
-}
