--- src/app/api/dashboard/route.ts
@@ -1,38 +1,0 @@
-import { promises as fs } from "fs";
-import path from "path";
-import { Book } from "@/app/types/book";
-
-export async function GET() {
-  try {
-    const filePath = path.join(process.cwd(), "data", "books.json");
-    const data = await fs.readFile(filePath, "utf-8");
-    const books: Book[] = JSON.parse(data);
-
-    const aberto = books.filter((b: any) => b.status === "aberto").length;
-    const fechado = books.filter((b: any) => b.status === "fechado").length;
-    const finalizados = books.filter(
-      (b: any) => b.status === "finalizado"
-    ).length;
-    const paginasLidas = books
-      .filter((b: any) => b.status === "finalizado")
-      .reduce((acc: number, b: any) => acc + (b.paginas || 0), 0);
-
-    return Response.json({
-      totalLivrosRegistrados: books.length,
-      livrosNaoIniciados: fechado,
-      livrosAbertos: aberto,
-      livrosFinalizados: finalizados,
-      totalPaginasLidas: paginasLidas,
-
-    });
-  } catch (error: any) {
-    return Response.json(
-      {
-        error: "Erro ao gerar estatísticas",
-        details:
-          "Revise o arquivo data/books.json. Talvez ele esteja corrompido ou vazio.",
-      },
-      { status: 500 }
-    );
-  }
-}
