--- src/app/layout.tsx
@@ -1,50 +1,0 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Readfy",
-  description: "Readfy - seu aplicativo de leitura",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        <nav className="bg-white shadow-sm border-b">
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="flex justify-between h-16">
-              <div className="flex items-center">
-                <h1 className="text-xl font-bold text-gray-900">📚 Readfy</h1>
-              </div>
-              <div className="flex items-center space-x-4">
-                <a href="/" className="text-gray-600 hover:text-gray-900">Início</a>
-                <a href="/dashboard" className="text-gray-600 hover:text-gray-900">Dashboard</a>
-                <a href="/book/register" className="bg-blue-500 text-white px-4 py-2 rounded-lg hover:bg-blue-600">
-                  + Novo Livro
-                </a>
-              </div>
-            </div>
-          </div>
-        </nav>
-        <main>{children}</main>
-      </body>
-    </html>
-  );
-}
